[CBRD-23903] Set tuple_count as INT64 type in list file (#2865)
http://jira.cubrid.org/browse/CBRD-23903

All kinds of queries to count # of rows are gathered into list_file.tuple_count
1) list_file->tuple_cnt is changed to INT64
- db_query_count(), db_query_seek_tuple(), insert/delete/update routines in execute_statement.c take list_file->tuple_cnt value. therefore they are changed.
- packing/unpacking at object_representation.c is changed. (aligning is considered)
2) stats that counting # of rows/keys in scan_stats are changed to print correctly for `;trace on`

+++ src/query/query_aggregate.cpp
@@ -2898,7 +2898,7 @@ qdata_aggregate_interpolation (cubthread::entry *thread_p, cubxasl::aggregate_li
 			       QFILE_LIST_SCAN_ID *scan_id)
 {
   int error = NO_ERROR;
-  int tuple_count;
+  INT64 tuple_count;
   double row_num_d, f_row_num_d, c_row_num_d, percentile_d;
   FUNC_TYPE function;
   double cur_group_percentile;
